backend/requirements.txt
Django>=5.0,<5.1
channels>=4.0,<4.1
channels-redis>=4.1,<4.2
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
psycopg2-binary>=2.9,<3.0
redis>=5.0,<5.1
daphne>=4.0,<4.1
python-dotenv>=1.0,<1.1
drf-spectacular>=0.27,<0.28
